refactor(usage): extract shared primitives and reorganize into subdirs
